refactor: move `GROUPS` and `register_groups()` to `crate::lints`, then delete `crate::groups`
