+++ bevy_lint/src/lints/mod.rs
@@ -20,7 +20,7 @@
 //! - [`restriction`]
 //! - [`nursery`]
 
-use crate::lint::BevyLint;
+use crate::lint::{BevyLint, LintGroup};
 use rustc_lint::{Lint, LintStore};
 
 mod cargo;
@@ -34,7 +34,11 @@ pub mod restriction;
 pub mod style;
 pub mod suspicious;
 
-pub(crate) static LINTS: &[&BevyLint] = &[
+/// A list of all [`BevyLint`]s.
+///
+/// If a group is not in this list, it will not be registered in [`register_lints()`].
+static LINTS: &[&BevyLint] = &[
+    // This list should be sorted alphabetically based on the lint's name, not group.
     pedantic::borrowed_reborrowable::BORROWED_REBORROWABLE,
     nursery::duplicate_bevy_dependencies::DUPLICATE_BEVY_DEPENDENCIES,
     suspicious::insert_event_resource::INSERT_EVENT_RESOURCE,
@@ -47,11 +51,28 @@ pub(crate) static LINTS: &[&BevyLint] = &[
     nursery::zst_query::ZST_QUERY,
 ];
 
+/// A list of all [`LintGroup`]s.
+///
+/// If a group is not in this list, it will not be registered in [`register_groups()`].
+static GROUPS: &[&LintGroup] = &[
+    // This list should be sorted alphabetically.
+    complexity::COMPLEXITY,
+    correctness::CORRECTNESS,
+    nursery::NURSERY,
+    pedantic::PEDANTIC,
+    performance::PERFORMANCE,
+    restriction::RESTRICTION,
+    style::STYLE,
+    suspicious::SUSPICIOUS,
+];
+
+/// Registers all [`BevyLint`]s in [`LINTS`] with the [`LintStore`].
 pub(crate) fn register_lints(store: &mut LintStore) {
     let lints: Vec<&Lint> = LINTS.iter().map(|x| x.lint).collect();
     store.register_lints(&lints);
 }
 
+/// Registers all lint passes with the [`LintStore`].
 pub(crate) fn register_passes(store: &mut LintStore) {
     store.register_late_pass(|_| {
         Box::new(pedantic::borrowed_reborrowable::BorrowedReborrowable::default())
@@ -77,4 +98,21 @@ pub(crate) fn register_passes(store: &mut LintStore) {
         Box::new(style::plugin_not_ending_in_plugin::PluginNotEndingInPlugin::default())
     });
     store.register_late_pass(|_| Box::new(nursery::zst_query::ZstQuery::default()));
+}
+
+/// Registers all [`LintGroup`]s in [`GROUPS`] with the [`LintStore`].
+pub(crate) fn register_groups(store: &mut LintStore) {
+    for &group in GROUPS {
+        let lints = LINTS
+            .iter()
+            .copied()
+            // Only select lints of this specified group.
+            .filter(|l| l.group == group)
+            // Convert the lints into their `LintId`s.
+            .map(BevyLint::id)
+            // Collect into a `Vec`.
+            .collect();
+
+        store.register_group(true, group.name, None, lints);
+    }
 }
